Refactoring: settings
