+++ src/pages/settings/visualsearch/IndexingInProgress/IndexingInProgress.tsx
